+++ src/games/stendhal/common/CollisionDetection.java
@@ -75,7 +75,7 @@ public class CollisionDetection
         {
         if(j>=0 && j<height && i>=0 && i<width)
           {
-          if(j==(int)y && i==(int)x)
+          if(j==y && i==x)
             {
             System.out.print("O");
             }
